Drop and recreate exp and iat columns as IntegerField

+++ sbs_project/landing/migrations/0004_alter_authdata_exp_alter_authdata_iat.py
@@ -36,11 +36,11 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='authdata',
             name='exp',
-            field=models.IntegerField(blank=True, null=True),
+            field=models.BigIntegerField(blank=True, null=True),
         ),
         migrations.AddField(
             model_name='authdata',
             name='iat',
-            field=models.IntegerField(blank=True, null=True),
+            field=models.BigIntegerField(blank=True, null=True),
         ),
     ]
